+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * The `api::projects::issues::Issues` endpoint's `milestone` field was
     changed to match the actual API exposed by GitLab (with `None` and `Any`
     options).
+  * The `api::projects::pipelines::PipelineVariables` endpoint is now pageable.
 
 # v0.1210.2